+++ SlimHUD/AppDelegate.swift
@@ -13,9 +13,8 @@ import Sparkle
 
 @NSApplicationMain
 class AppDelegate: NSWindowController, NSApplicationDelegate {
-    let statusItem = NSStatusBar.system.statusItem(withLength: NSStatusItem.squareLength)
-
     var settingsManager: SettingsManager = SettingsManager.getInstance()
+    var settingsWindowController: SettingsWindowController?
 
     var volumeHud = Hud()
     var brightnessHud = Hud()
@@ -25,17 +24,10 @@ class AppDelegate: NSWindowController, NSApplicationDelegate {
     lazy var displayer = Displayer(positionManager: positionManager, volumeHud: volumeHud, brightnessHud: brightnessHud, keyboardHud: keyboardHud)
     lazy var changesObserver = ChangesObserver(positionManager: positionManager, displayer: displayer)
 
+    weak var settingsViewTabsManager: TabsManager?
+
     override func awakeFromNib() {
         super.awakeFromNib()
-
-        // menu bar
-        statusItem.menu = statusMenu
-
-        if let button = statusItem.button {
-            button.image = IconManager.getStatusIcon()
-            button.image?.isTemplate = true
-        }
-
         displayer.updateAllAttributes()
     }
 
@@ -54,8 +46,72 @@ class AppDelegate: NSWindowController, NSApplicationDelegate {
         }
 
         OSDUIManager.stop()
+
+        if CommandLine.arguments.contains("showSettingsAtLaunch") {
+            showSettingsWindow()
+        }
     }
 
-    @IBOutlet weak var statusMenu: NSMenu!
+    func applicationDidBecomeActive(_ notification: Notification) {
+        showSettingsWindow()
+    }
 
+    func showSettingsWindow() {
+        if let settingsWindowController = settingsWindowController {
+            settingsWindowController.showWindow(self)
+        } else {
+            if let windowController = NSStoryboard(name: "Settings", bundle: nil).instantiateInitialController() as? SettingsWindowController {
+                settingsWindowController = windowController
+                windowController.showWindow(self)
+            }
+        }
+        NSApplication.shared.setActivationPolicy(.regular)
+        NSApp.activate(ignoringOtherApps: true)
+    }
+
+    @IBAction func quitCliked(_ sender: Any) {
+        if isSomeWindowVisible() {
+            let alertResponse = showAlert(question: "SlimHUD will continue to show HUDs in the background",
+                                          text: "If you want to quit, click \"Quit now\"",
+                                          buttonsTitle: ["OK", "Quit now"])
+            if alertResponse == NSApplication.ModalResponse.alertSecondButtonReturn {
+                quit()
+            }
+            closeAllWindows()
+            NSApplication.shared.setActivationPolicy(.accessory)
+        } else {
+            quit()
+        }
+    }
+
+    @IBAction func openGeneralTab(_ sender: Any) {
+        settingsViewTabsManager?.selectItem(index: 0)
+    }
+
+    @IBAction func openDesignTab(_ sender: Any) {
+        settingsViewTabsManager?.selectItem(index: 1)
+    }
+
+    @IBAction func openStyleTab(_ sender: Any) {
+        settingsViewTabsManager?.selectItem(index: 2)
+    }
+
+    @IBAction func openAboutTab(_ sender: Any) {
+        settingsViewTabsManager?.selectItem(index: 3)
+    }
+
+    private func closeAllWindows() {
+        settingsWindowController?.close()
+    }
+
+    private func quit() {
+        settingsManager.saveAllItems()
+        OSDUIManager.start()
+        exit(0)
+    }
+
+    private func isSomeWindowVisible() -> Bool {
+        return (settingsWindowController?.window?.isVisible ?? false) &&
+            NSApplication.shared.activationPolicy() != .accessory
+    }
 }