+++ SlimHUD/Controllers/ HudsControllerInterface.swift
@@ -10,8 +10,9 @@ import Cocoa
 protocol HudsControllerInterface: AnyObject {
     func updateShadows(enabled: Bool)
     func hideIcon(isHidden: Bool)
-    func setupDefaultBarsColors()
-    func setBackgroundColor(color: NSColor)
+    func setVolumeBackgroundColor(color: NSColor)
+    func setBrightnessBackgroundColor(color: NSColor)
+    func setKeyboardBackgroundColor(color: NSColor)
     func setVolumeEnabledColor(color: NSColor)
     func setVolumeDisabledColor(color: NSColor)
     func setBrightnessColor(color: NSColor)
@@ -25,7 +26,5 @@ protocol HudsControllerInterface: AnyObject {
     func setBrightnessIconsTint(_ color: NSColor)
     @available(OSX 10.14, *)
     func setKeyboardIconsTint(_ color: NSColor)
-    @available(OSX 10.14, *)
-    func setupDefaultIconsColors()
     var positionManager: PositionManager { get }
 }