+++ SlimHUD/Services/Displayer.swift
@@ -15,6 +15,8 @@ class Displayer: HudsControllerInterface {
     private var brightnessHud: Hud
     private var keyboardHud: Hud
 
+    public var temporarelyEnableAllBars = false
+
     init(positionManager: PositionManager, volumeHud: Hud, brightnessHud: Hud, keyboardHud: Hud) {
         self.positionManager = positionManager
         self.volumeHud = volumeHud
@@ -27,7 +29,7 @@ class Displayer: HudsControllerInterface {
     }
 
     func showVolumeHUD() {
-        if !settingsManager.enabledBars.volumeBar {
+        if !(settingsManager.enabledBars.volumeBar || temporarelyEnableAllBars) {
             return
         }
         let isMuted = VolumeManager.isMuted()
@@ -44,7 +46,7 @@ class Displayer: HudsControllerInterface {
     }
 
     func showBrightnessHUD() {
-        if !settingsManager.enabledBars.brightnessBar {
+        if !(settingsManager.enabledBars.brightnessBar || temporarelyEnableAllBars) {
             return
         }
         // if the function is being called because the key has been pressed, the display's brightness
@@ -55,6 +57,7 @@ class Displayer: HudsControllerInterface {
                 self.brightnessHud.setProgress(progress: progress)
                 self.brightnessHud.setIconImage(icon: IconManager.getBrightnessIcon(for: progress))
             } catch {
+                self.brightnessHud.setProgress(progress: 0.5)
                 NSLog("Failed to retrieve display brightness. See https://github.com/AlexPerathoner/SlimHUD/issues/60")
             }
         }
@@ -67,7 +70,7 @@ class Displayer: HudsControllerInterface {
         brightnessHud.dismiss(delay: 1.5)
     }
     func showKeyboardHUD() {
-        if !settingsManager.enabledBars.keyboardBar {
+        if !(settingsManager.enabledBars.keyboardBar || temporarelyEnableAllBars) {
             return
         }
         // if the function is being called because the key has been pressed, the keyboard's brightness
@@ -78,6 +81,7 @@ class Displayer: HudsControllerInterface {
                 self.keyboardHud.setProgress(progress: progress)
                 self.keyboardHud.setIconImage(icon: IconManager.getKeyboardIcon(for: progress))
             } catch {
+                self.keyboardHud.setProgress(progress: 0.5)
                 NSLog("Failed to retrieve display brightness. See https://github.com/AlexPerathoner/SlimHUD/issues/60")
             }
         }
@@ -110,24 +114,13 @@ class Displayer: HudsControllerInterface {
         keyboardHud.hideIcon(isHidden: isHidden)
     }
 
-    func setupDefaultBarsColors() {
-        setVolumeEnabledColor(color: DefaultColors.Blue)
-        setVolumeDisabledColor(color: DefaultColors.Gray)
-        setBrightnessColor(color: DefaultColors.Yellow)
-        setKeyboardColor(color: DefaultColors.Azure)
-        setBackgroundColor(color: DefaultColors.DarkGray)
-    }
-
-    @available(OSX 10.14, *)
-    func setupDefaultIconsColors() {
-        setVolumeIconsTint(.white)
-        setBrightnessIconsTint(.white)
-        setKeyboardIconsTint(.white)
-    }
-
-    func setBackgroundColor(color: NSColor) {
+    func setVolumeBackgroundColor(color: NSColor) {
         volumeHud.setBackgroundColor(color: color)
+    }
+    func setBrightnessBackgroundColor(color: NSColor) {
         brightnessHud.setBackgroundColor(color: color)
+    }
+    func setKeyboardBackgroundColor(color: NSColor) {
         keyboardHud.setBackgroundColor(color: color)
     }
     func setVolumeEnabledColor(color: NSColor) {
@@ -148,7 +141,9 @@ class Displayer: HudsControllerInterface {
         setThickness(thickness: CGFloat(settingsManager.barThickness))
         hideIcon(isHidden: !settingsManager.shouldShowIcons)
         updateShadows(enabled: settingsManager.shouldShowShadows)
-        setBackgroundColor(color: settingsManager.backgroundColor)
+        setVolumeBackgroundColor(color: settingsManager.volumeBackgroundColor)
+        setBrightnessBackgroundColor(color: settingsManager.brightnessBackgroundColor)
+        setKeyboardBackgroundColor(color: settingsManager.keyboardBackgroundColor)
         setVolumeEnabledColor(color: settingsManager.volumeEnabledColor)
         setVolumeDisabledColor(color: settingsManager.volumeDisabledColor)
         setBrightnessColor(color: settingsManager.brightnessColor)
@@ -160,7 +155,7 @@ class Displayer: HudsControllerInterface {
             setKeyboardIconsTint(settingsManager.keyboardIconColor)
         }
     }
-    
+
     func setAnimationStyle(animationStyle: AnimationStyle) {
         volumeHud.setAnimationStyle(animationStyle)
         brightnessHud.setAnimationStyle(animationStyle)