+++ SlimHUD/Services/HudAnimator.swift
@@ -11,7 +11,7 @@ class HudAnimator {
     private static let GrowShrinkFactor: CGFloat = 1.6
     private static let GrowFactorComplementary: CGFloat = (1-GrowShrinkFactor) / 2
     private static let ShrinkFactorComplementary: CGFloat = (1-1/GrowShrinkFactor) / 2
-    
+
     public static func popIn(barView: BarView) {
         barView.alphaValue = 1
     }
@@ -35,7 +35,7 @@ class HudAnimator {
             barView.animator().setFrameOrigin(getAnimationFrameOrigin(originPosition: originPosition, screenEdge: screenEdge))
         }, completionHandler: completion)
     }
-    
+
     public static func fadeIn(barView: BarView) {
         barView.alphaValue = 0
         NSAnimationContext.runAnimationGroup({ (context) in
@@ -50,7 +50,7 @@ class HudAnimator {
             barView.animator().alphaValue = 0
         }, completionHandler: completion)
     }
-    
+
     public static func growIn(barView: BarView) {
         barView.alphaValue = 0
         let originalBounds = barView.bounds
@@ -58,7 +58,7 @@ class HudAnimator {
                                             y: originalBounds.height * GrowFactorComplementary,
                                             width: originalBounds.width * GrowShrinkFactor,
                                             height: originalBounds.height * GrowShrinkFactor)
-        
+
         NSAnimationContext.runAnimationGroup({ (context) in
             context.duration = Constants.Animation.Duration
             barView.animator().alphaValue = 1
@@ -70,7 +70,7 @@ class HudAnimator {
     public static func growOut(barView: BarView, completion: @escaping (() -> Void)) {
         barView.alphaValue = 1
         let originalBounds = barView.bounds
-        
+
         NSAnimationContext.runAnimationGroup({ (context) in
             context.duration = Constants.Animation.Duration
             barView.animator().alphaValue = 0
@@ -87,7 +87,10 @@ class HudAnimator {
     public static func shrinkIn(barView: BarView) {
         barView.alphaValue = 0
         let originalBounds = barView.bounds
-        barView.bounds = NSRect(x: originalBounds.width * ShrinkFactorComplementary, y: originalBounds.height * ShrinkFactorComplementary, width: originalBounds.width / GrowShrinkFactor, height: originalBounds.height / GrowShrinkFactor)
+        barView.bounds = NSRect(x: originalBounds.width * ShrinkFactorComplementary,
+                                y: originalBounds.height * ShrinkFactorComplementary,
+                                width: originalBounds.width / GrowShrinkFactor,
+                                height: originalBounds.height / GrowShrinkFactor)
         NSAnimationContext.runAnimationGroup({ (context) in
             context.duration = Constants.Animation.Duration
             barView.animator().alphaValue = 1
@@ -102,22 +105,25 @@ class HudAnimator {
         NSAnimationContext.runAnimationGroup({ (context) in
             context.duration = Constants.Animation.Duration
             barView.animator().alphaValue = 0
-            barView.animator().bounds = NSRect(x: originalBounds.width * ShrinkFactorComplementary, y: originalBounds.height * ShrinkFactorComplementary, width: originalBounds.width / GrowShrinkFactor, height: originalBounds.height / GrowShrinkFactor)
+            barView.animator().bounds = NSRect(x: originalBounds.width * ShrinkFactorComplementary,
+                                               y: originalBounds.height * ShrinkFactorComplementary,
+                                               width: originalBounds.width / GrowShrinkFactor,
+                                               height: originalBounds.height / GrowShrinkFactor)
         }, completionHandler: {
             barView.bounds = originalBounds
             completion()
         })
     }
-    
+
     public static func sideGrowIn(barView: BarView, originPosition: CGPoint) {
         barView.alphaValue = 0
-        
+
         let originalBounds = barView.bounds
         barView.bounds = NSRect(x: originalBounds.width * GrowFactorComplementary,
                                             y: originalBounds.height * GrowFactorComplementary,
                                             width: originalBounds.width * GrowShrinkFactor,
                                             height: originalBounds.height * GrowShrinkFactor)
-        
+
         NSAnimationContext.runAnimationGroup({ (context) in
             context.duration = Constants.Animation.Duration
             barView.animator().alphaValue = 1
@@ -130,7 +136,7 @@ class HudAnimator {
     public static func sideGrowOut(barView: BarView, originPosition: CGPoint, screenEdge: Position, completion: @escaping (() -> Void)) {
         barView.alphaValue = 1
         let originalBounds = barView.bounds
-        
+
         NSAnimationContext.runAnimationGroup({ (context) in
             context.duration = Constants.Animation.Duration
             barView.animator().alphaValue = 0
@@ -144,7 +150,7 @@ class HudAnimator {
             completion()
         })
     }
-    
+
     public static func getAnimationFrameOrigin(originPosition: CGPoint, screenEdge: Position) -> CGPoint {
         switch screenEdge {
         case .left: