+++ SlimHUD/Services/SettingsManager.swift
@@ -11,7 +11,9 @@ class SettingsManager {
 
     // MARK: - Keys
 
-    private static let BackgroundColorKey = "backgroundColor"
+    private static let VolumeBackgroundColorKey = "volumeBackgroundColor"
+    private static let BrightnessBackgroundColorKey = "brightnessBackgroundColor"
+    private static let KeyboardBackgroundColorKey = "keyboardBackgroundColor"
     private static let VolumeEnabledColorKey = "volumeEnabledColor"
     private static let VolumeDisabledColorKey = "volumeDisabledColor"
     private static let BrightnessColorKey = "brightnessColor"
@@ -28,13 +30,22 @@ class SettingsManager {
     private static let PositionKey = "position"
     private static let EnabledBarsKey = "enabledBars"
     private static let MarginKey = "marginValue"
-    private static let ShowQuitAlertKey = "showQuitAlert"
     private static let FlatBarKey = "flatBar"
 
     // MARK: - Bars colors
-    var backgroundColor: NSColor {
+    var volumeBackgroundColor: NSColor {
         didSet {
-            UserDefaultsManager.setItem(backgroundColor, for: SettingsManager.BackgroundColorKey)
+            UserDefaultsManager.setItem(volumeBackgroundColor, for: SettingsManager.VolumeBackgroundColorKey)
+        }
+    }
+    var brightnessBackgroundColor: NSColor {
+        didSet {
+            UserDefaultsManager.setItem(brightnessBackgroundColor, for: SettingsManager.BrightnessBackgroundColorKey)
+        }
+    }
+    var keyboardBackgroundColor: NSColor {
+        didSet {
+            UserDefaultsManager.setItem(keyboardBackgroundColor, for: SettingsManager.KeyboardBackgroundColorKey)
         }
     }
     var volumeEnabledColor: NSColor {
@@ -134,12 +145,6 @@ class SettingsManager {
         }
     }
 
-    var showQuitAlert: Bool {
-        didSet {
-            UserDefaults.standard.set(showQuitAlert, forKey: SettingsManager.ShowQuitAlertKey)
-        }
-    }
-
     // MARK: - Class methods
 
     static private let singletonSettingsController = SettingsManager()
@@ -148,7 +153,9 @@ class SettingsManager {
     }
 
     private init() {
-        backgroundColor = UserDefaultsManager.getItem(for: SettingsManager.BackgroundColorKey, defaultValue: DefaultColors.DarkGray)
+        volumeBackgroundColor = UserDefaultsManager.getItem(for: SettingsManager.VolumeBackgroundColorKey, defaultValue: DefaultColors.DarkGray)
+        brightnessBackgroundColor = UserDefaultsManager.getItem(for: SettingsManager.BrightnessBackgroundColorKey, defaultValue: DefaultColors.DarkGray)
+        keyboardBackgroundColor = UserDefaultsManager.getItem(for: SettingsManager.KeyboardBackgroundColorKey, defaultValue: DefaultColors.DarkGray)
         volumeEnabledColor = UserDefaultsManager.getItem(for: SettingsManager.VolumeEnabledColorKey, defaultValue: DefaultColors.Blue)
         volumeDisabledColor = UserDefaultsManager.getItem(for: SettingsManager.VolumeDisabledColorKey, defaultValue: DefaultColors.Gray)
         brightnessColor = UserDefaultsManager.getItem(for: SettingsManager.BrightnessColorKey, defaultValue: DefaultColors.Yellow)
@@ -174,16 +181,13 @@ class SettingsManager {
              enabledBarsRaw[EnabledBars.KeyboardBarIndex])
         enabledBars = EnabledBars(volumeBar: volumeBarEnabled, brightnessBar: brightnessBarEnabled, keyboardBar: keyboardBarEnabled)
         marginValue = UserDefaultsManager.getInt(for: SettingsManager.MarginKey, defaultValue: 10)
-        if CommandLine.arguments.contains("showQuitAlert") {
-            let indexOfValue = CommandLine.arguments.firstIndex(of: "showQuitAlert")! + 1
-            UserDefaults.standard.set(CommandLine.arguments[indexOfValue], forKey: SettingsManager.ShowQuitAlertKey)
-        }
-        showQuitAlert = UserDefaultsManager.getBool(for: SettingsManager.ShowQuitAlertKey, defaultValue: true)
         flatBar = UserDefaultsManager.getBool(for: SettingsManager.FlatBarKey, defaultValue: false)
     }
 
     func resetDefaultBarsColors() {
-        backgroundColor = DefaultColors.DarkGray
+        volumeBackgroundColor = DefaultColors.DarkGray
+        brightnessBackgroundColor = DefaultColors.DarkGray
+        keyboardBackgroundColor = DefaultColors.DarkGray
         volumeEnabledColor = DefaultColors.Blue
         volumeDisabledColor = DefaultColors.Gray
         brightnessColor = DefaultColors.Yellow
@@ -197,7 +201,9 @@ class SettingsManager {
     }
 
     func saveAllItems() {
-        UserDefaultsManager.setItem(backgroundColor, for: SettingsManager.BackgroundColorKey)
+        UserDefaultsManager.setItem(volumeBackgroundColor, for: SettingsManager.VolumeBackgroundColorKey)
+        UserDefaultsManager.setItem(brightnessBackgroundColor, for: SettingsManager.BrightnessBackgroundColorKey)
+        UserDefaultsManager.setItem(keyboardBackgroundColor, for: SettingsManager.KeyboardBackgroundColorKey)
         UserDefaultsManager.setItem(volumeEnabledColor, for: SettingsManager.VolumeEnabledColorKey)
         UserDefaultsManager.setItem(volumeDisabledColor, for: SettingsManager.VolumeDisabledColorKey)
         UserDefaultsManager.setItem(brightnessColor, for: SettingsManager.BrightnessColorKey)