+++ SlimHUD/Support/AnimationStyle.swift
@@ -8,23 +8,23 @@
 import Foundation
 
 enum AnimationStyle: String {
-    case None
-    case Slide
-    case PopInFadeOut = "Pop in + fade out"
-    case Fade
-    case Grow
-    case Shrink
-    case SideGrow = "Side grow"
-    
-    private static let DefaultValue = AnimationStyle.Slide
-    
+    case none = "None"
+    case slide = "Slide"
+    case popInFadeOut = "Pop in + fade out"
+    case fade = "Fade"
+    case grow = "Grow"
+    case shrink = "Shrink"
+    case sideGrow = "Side grow"
+
+    private static let DefaultValue = AnimationStyle.slide
+
     init(from rawValue: String?) {
         self = AnimationStyle(rawValue: rawValue ?? "") ?? AnimationStyle.DefaultValue
     }
-    
+
     func requiresInMovement() -> Bool {
         switch self {
-        case .Slide, .SideGrow:
+        case .slide, .sideGrow:
             return true
         default:
             return false