+++ SlimHUDUITests/SettingsUITest.swift
@@ -6,9 +6,9 @@
 //  Copyright © 2022 Alex Perathoner. All rights reserved.
 //
 
- import XCTest
+import XCTest
 
- final class SettingsUITest: SparkleUITests {
+final class SettingsUITest: SparkleUITests {
     override func setUpWithError() throws {
         continueAfterFailure = false
     }
@@ -16,23 +16,22 @@
     func testOpenSettingsWindow() throws {
         let app = XCUIApplication()
         app.launch()
-        let settingsWindow = openSettingsWindow(app)
 
-        addScreenshot(window: settingsWindow, name: "Settings window")
+        XCTAssertTrue(app.windows.count == 0)
 
-        usleep(500000)
-        // preview HUD will also appear, so the windows count should be more than 1
-        XCTAssertEqual(app.windows.count, 2)
+        app.activate()
+
+        let settingsWindow = app.windows.matching(identifier: "SlimHUD").firstMatch
+        XCTAssert(settingsWindow.waitForExistence(timeout: 5))
+        addScreenshot(window: settingsWindow, name: "Settings")
     }
 
     func testCloseWindow() throws {
         let app = XCUIApplication()
-        app.showCmdQAlert(false)
         app.launch()
+        app.activate()
 
-        // try closing with cmd + w
-        let statusItem = UITestsUtils.getStatusItem(app: app)
-        var settingsWindow = openSettingsWindow(app)
+        var settingsWindow = app.windows.matching(identifier: "SlimHUD").firstMatch
 
         settingsWindow.typeKey("w", modifierFlags: .command)
 
@@ -41,32 +40,14 @@
         // relaunching as the app is now in background and doesn't accept test interaction
 
         app.launch()
+        app.activate()
 
         // try closing with cmd + q
-        settingsWindow = openSettingsWindow(app)
+        settingsWindow = app.windows.matching(identifier: "SlimHUD").firstMatch
 
         settingsWindow.typeKey("q", modifierFlags: .command)
+        app.dialogs["alert"].buttons["OK"].click()
 
         XCTAssertFalse(settingsWindow.isHittable)
-        // app should still be running
-        XCTAssertTrue(statusItem.exists)
-    }
-
-     private func openSettingsWindow(_ app: XCUIApplication) -> XCUIElement {
-
-        let statusItem = SparkleUITests.getStatusItem(app: app)
-
-        let preferencesMenuItem = statusItem.menuItems["Settings..."]
-
-        let settingsWindow = app.windows.matching(identifier: "Settings").firstMatch
-
-        statusItem.click()
-
-        XCTAssert(preferencesMenuItem.waitForExistence(timeout: 5))
-        preferencesMenuItem.click()
-
-        XCTAssert(settingsWindow.waitForExistence(timeout: 5))
-
-        return settingsWindow
     }
  }