+++ SlimHUDUITests/Utils/UITestsUtils.swift
@@ -5,20 +5,13 @@
 //  Created by Alex Perathoner on 05/01/23.
 //
 
-import XCTest
+ import XCTest
 
-class UITestsUtils: XCTestCase {
-    static public func getStatusItem(app: XCUIApplication) -> XCUIElement {
-        let menuBarsQuery = app.menuBars
-        let statusItem = menuBarsQuery.children(matching: .statusItem).element(boundBy: 0)
-        XCTAssert(statusItem.waitForExistence(timeout: 5))
-        return statusItem
-    }
-
+ class UITestsUtils: XCTestCase {
     public func addScreenshot(window: XCUIElement, name: String) {
         let attachment = XCTAttachment(screenshot: window.screenshot())
         attachment.name = name
         attachment.lifetime = .keepAlways
         add(attachment)
     }
-}
+ }