+++ SlimHUDUITests/Utils/XCUIApplicationExtension.swift
@@ -11,7 +11,7 @@ extension XCUIApplication {
     func shouldContinuouslyCheck() {
         launchArguments += ["shouldContinuouslyCheck"]
     }
-    func showCmdQAlert(_ showCmdQAlert: Bool) {
-        launchArguments += ["showQuitAlert", showCmdQAlert.description]
+    func shouldShowSettingsAtLaunch() {
+        launchArguments += ["showSettingsAtLaunch"]
     }
 }