+++ Release_Notes.md
@@ -1,6 +1,9 @@
 # 1.5.0 - New look
 * Icons change depending on value
 * Add multiple in-out animation styles
+* New look for the settings window
+
+@kaydenanderson
 
 # 1.4.2 - Improved non-continous checks
 * Show system HUDs after SlimHUD quit